UDENG-418: allow access to pipewire socket
This MR is a first proposal to allow access to the pipewire
socket from snapped applications. To avoid security issues,
it only allows to access the sockets located at the folder
belonging to the snap that owns the slot.

With this patch, audio works fine in Core Desktop, and also
screen sharing.