interfaces/home: Autoconnect home for core desktop
This adds a new on-core-desktop option for interfaces and autoconnects home if
this is available.

Note that it checks before the classic check as otherwise the classic check will
fail for home and core desktop is never checked. This needs to be resolved
before this change can land in snapd.